update: cartitem id in repository

+++ src/main/java/com/kavindu/commercehub/Cart/repositories/CartRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface CartRepository extends JpaRepository<CartItem, Long> {
+public interface CartRepository extends JpaRepository<CartItem, Integer> {
     List<CartItem> findByCustomerId(UUID customerId);
     Optional<CartItem> findByCustomerIdAndProductId(Long customerId, Long productId);
 }
